feat(plz): move pkg/resources/plz to pkg/plz
